+++ spec/avro/dsl/test/recursive.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 record :recursive, namespace: :test do
   required :s, :string
   optional :child, :recursive