+++ spec/avro/dsl/test/with_map.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 record :with_map, namespace: :test do
   required :pairs, :map, values: :int, default: { a: 1 }
 end