+++ spec/avro/dsl/test/with_union.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 record :with_union, namespace: :test do
   optional :s, :string
 end