+++ spec/avromatic_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 describe Avromatic do
   it "has a version number" do
     expect(Avromatic::VERSION).not_to be_nil