+++ lib/avromatic/io.rb
@@ -1,7 +1,9 @@
+# frozen_string_literal: true
+
 module Avromatic
   module IO
-    UNION_MEMBER_INDEX = '__avromatic_member_index'.freeze
-    ENCODING_PROVIDER = '__avromatic_encoding_provider'.freeze
+    UNION_MEMBER_INDEX = '__avromatic_member_index'
+    ENCODING_PROVIDER = '__avromatic_encoding_provider'
   end
 end