+++ lib/avromatic/model/value_object.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Avromatic
   module Model
     # This module is used to override the comparisons defined by