+++ spec/avro/dsl/test/encode_value.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 record :encode_value, namespace: :test do
   required :str1, :string, default: 'X'
   required :str2, :string, default: 'Y'