+++ spec/avro/dsl/test/key_conflict.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 record :key_conflict, namespace: 'test' do
   required :id, :bytes
   required :b, :string