+++ spec/avro/dsl/test/key_overlap.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 record :key_overlap, namespace: 'test' do
   required :id, :long
   required :b, :string