+++ spec/avro/dsl/test/key_with_optional.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 record :key_with_optional, namespace: :test do
   required :id, :long
   optional :name, :string