+++ spec/avro/dsl/test/with_array.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 record :with_array, namespace: :test do
   required :names, :array, items: :string, default: ['first']
 end