+++ spec/avromatic/model/messaging_serialization_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'avro/builder'
 
 describe Avromatic::Model::MessagingSerialization do