+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   but errors will now appear under the `:base` key.  
 - Support for custom types in unions with more than one non-null type.
 - Drop support for Ruby < 2.3 and Rails < 5.0.
+- Call `super()` in model constructor.
 
 ## v1.0.0
 - No changes.